fix: sorting tests.
